Aggiunta funzioni in PropostaRepository e cambio definizione di prezzoProposta da BigDecimal a Double

+++ backend/src/main/java/it/autoflow/proposal/entity/Proposta.java
@@ -7,7 +7,7 @@ import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.Setter;
 
-import java.math.BigDecimal;
+
 import java.time.LocalDate;
 
 @Getter @Setter
@@ -29,7 +29,7 @@ public class Proposta {
     private Configurazione configurazione;
 
     @Column(nullable = false)
-    private BigDecimal prezzoProposta;
+    private Double prezzoProposta;
 
     @Enumerated(EnumType.STRING)
     @Column(nullable = false)
